Refactor `mq` module into multiple submodules for improved organization
