+++ neon_data_models/models/api/mq/users.py
@@ -24,12 +24,10 @@
 # NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 # SOFTWARE,  EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-from typing import Literal, Optional, Annotated, Union, Dict, List
-
+from typing import Literal, Optional, Annotated, Union
 from pydantic import Field, TypeAdapter, model_validator
 
 from neon_data_models.models.api.jwt import HanaToken
-from neon_data_models.models.api.llm import LLMRequest, LLMPersona
 from neon_data_models.models.base.contexts import MQContext
 from neon_data_models.models.user.database import User
 
@@ -105,44 +103,6 @@ class UserDbRequest:
         return cls.ta.validate_python(kwargs)
 
 
-class LLMProposeRequest(MQContext, LLMRequest):
-    model: Optional[str] = Field(
-        default=None,
-        description="MQ implementation defines `model` as optional because the "
-                    "queue defines the requested model in most cases.")
-    persona: Optional[LLMPersona] = Field(
-        default=None,
-        description="MQ implementation defines `persona` as an optional "
-                    "parameter, with default behavior hard-coded into each "
-                    "LLM module.")
-
-
-class LLMProposeResponse(MQContext):
-    response: str = Field(description="LLM response to the prompt")
-
-
-class LLMDiscussRequest(LLMProposeRequest):
-    options: Dict[str, str] = Field(
-        description="Mapping of participant name to response to be discussed.")
-
-
-class LLMDiscussResponse(MQContext):
-    opinion: str = Field(description="LLM response to the available options.")
-
-
-class LLMVoteRequest(LLMProposeRequest):
-    responses: List[str] = Field(
-        description="List of responses to choose from.")
-
-
-class LLMVoteResponse(MQContext):
-    sorted_answer_indexes: List[int] = Field(
-        description="Indices of `responses` ordered high to low by preference.")
-
-
 __all__ = [CreateUserRequest.__name__, ReadUserRequest.__name__,
            UpdateUserRequest.__name__, DeleteUserRequest.__name__,
-           UserDbRequest.__name__, LLMProposeRequest.__name__,
-           LLMProposeResponse.__name__, LLMDiscussRequest.__name__,
-           LLMDiscussResponse.__name__, LLMVoteRequest.__name__,
-           LLMVoteResponse.__name__]
+           UserDbRequest.__name__]
